working with pdbbind now

+++ deepchem_debug.py
@@ -1,15 +1,18 @@
-
-import torch
-
-
+import rdkit
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+import networkx as nx
+import mbuild as mb
+import parmed
 import torch
 from torch import nn
 import torch_geometric as tg
 from torch_geometric.loader import DataLoader
-from typing import List, Optional, Union
 import torch.utils.data
 import shutil
 import sys
+import rdkit 
 sys.path.insert(1, '~/egnn_cof/models/egnn_clean')
 sys.path.insert(1, '~/egnn_cof')
 # from easydict import EasyDict as edict
@@ -20,79 +23,59 @@ from loader import *
 from megnn import *
 from utils import *
 
-def compute_cof_mean_mad(dataframe):
-    values = torch.Tensor(dataframe['COF'])
-    mean = torch.mean(values)
-    ma = torch.abs(values - mean)
-    mad = torch.mean(ma)
-    return mean, mad
-
 # clear the processed dataset
-try:
-	shutil.rmtree('./processed')
-except:
-	pass
-n_epochs  = 30
 
+try:
+    if sys.argv[1] == 'remove':
+        print('Removed processed dataset')
+        shutil.rmtree('./processed')
+except:
+    pass
+
+n_epochs  = 5
 device = torch.device("cpu")
 dtype = torch.float32
+dat = PdbBind_Dataset(root='.')
 
-dat = COF_Dataset(root='.')
-
-batch_size = 32
+batch_size = 4
 
 dat.shuffle()
-train_dataset = dat[:int(len(dat)*.8)]
-test_dataset = dat[int(len(dat)*.8):]
+train_dataset = dat[:4000]
+test_dataset = dat[4000:5000]
+valid_dataset = dat[5000:]
 train_loader = DataLoader(train_dataset, batch_size=batch_size, follow_batch=['x_s', 'x_t', 'positions_s', 'positions_t'], shuffle=True)
-test_loader = DataLoader(test_dataset, batch_size=batch_size, follow_batch=['x_s', 'x_t', 'positions_s', 'positions_t'], shuffle=False)
+test_loader = DataLoader(test_dataset, batch_size=batch_size, follow_batch=['x_s', 'x_t','positions_s', 'positions_t'], shuffle=False)
+valid_loader = DataLoader(valid_dataset, batch_size=batch_size, follow_batch=['x_s', 'x_t','positions_s', 'positions_t'], shuffle=False)
 
-def compute_cof_mean_mad(data):
-    values = torch.Tensor(data['COF'])
-    mean = torch.mean(values)
-    ma = torch.abs(values - mean)
-    mad = torch.mean(ma)
-    return mean, mad
+model = MEGNN(n_graphs=2, in_node_nf=dat.data.x_s.size(1), in_edge_nf=0, hidden_nf=128, device=device, n_layers=7, coords_weight=1.0,
+             attention=True, node_attr=1)
+# model = PairEGNN(in_node_nf=len(dat.elements), in_edge_nf=0, hidden_nf=128, device=device, n_layers=7, coords_weight=1.0,
+#              attention=True, node_attr=1)
 
-prop_mean, prop_mad = compute_cof_mean_mad(dat.dataframe)
-
-model = PairEGNN(in_node_nf=5, in_edge_nf=0, hidden_nf=128, device=device, n_layers=8, coords_weight=1.,
-             attention=False, node_attr=1)
-
-print(model)
-
-optimizer = optim.Adam(model.parameters(), lr=1e-3, weight_decay=1e-16)
-lr_scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(optimizer, n_epochs)
-criterion = nn.L1Loss()
+optimizer = optim.Adam(model.parameters())
+loss_func = nn.MSELoss()
 
 
-def train(epoch, loader):
-    epoch_loss = 0
-    lr_scheduler.step()
+def train():
     model.train()
-
-    for i, data in enumerate(loader):
+    epoch_loss = 0
+    for data in train_loader:  # Iterate in batches over the training dataset.
         dense_positions_s, atom_mask_s = tg.utils.to_dense_batch(data.positions_s, data.positions_s_batch)
         dense_positions_t, atom_mask_t = tg.utils.to_dense_batch(data.positions_t, data.positions_t_batch)
         batch_size_s, n_nodes_s, _ = dense_positions_s.size()
         batch_size_t, n_nodes_t, _ = dense_positions_t.size()
         atom_positions_s = dense_positions_s.view(batch_size_s * n_nodes_s, -1).to(device, dtype)
         atom_positions_t = dense_positions_t.view(batch_size_t * n_nodes_t, -1).to(device, dtype)
-
         edge_mask_s = atom_mask_s.unsqueeze(1) * atom_mask_s.unsqueeze(2)
-        #mask diagonal
         diag_mask = ~torch.eye(edge_mask_s.size(1), dtype=torch.bool).unsqueeze(0)
         edge_mask_s *= diag_mask
         edge_mask_s = edge_mask_s.view(batch_size_s * n_nodes_s * n_nodes_s, 1).to(device)
         edge_mask_t = atom_mask_t.unsqueeze(1) * atom_mask_t.unsqueeze(2)
-        #mask diagonal
         diag_mask = ~torch.eye(edge_mask_t.size(1), dtype=torch.bool).unsqueeze(0)
         edge_mask_t *= diag_mask
         edge_mask_t = edge_mask_t.view(batch_size_t * n_nodes_t * n_nodes_t, 1).to(device)
-
         atom_mask_s = atom_mask_s.view(batch_size_s * n_nodes_s, -1).to(device)
         atom_mask_t = atom_mask_t.view(batch_size_t * n_nodes_t, -1).to(device)
-
         one_hot_s, one_hot_s_mask = tg.utils.to_dense_batch(data.x_s, data.x_s_batch)
         one_hot_t, one_hot_s_mask = tg.utils.to_dense_batch(data.x_t, data.x_t_batch)
         one_hot_s = one_hot_s.view(batch_size_s * n_nodes_s, -1).to(device)
@@ -100,26 +83,25 @@ def train(epoch, loader):
         edges_s = get_adj_matrix(n_nodes_s, batch_size_s, device)
         edges_t = get_adj_matrix(n_nodes_t, batch_size_t, device)
         label = data.y.to(device, dtype)
-
-        pred = model(h0_s=one_hot_s, h0_t=one_hot_t, edges_s=edges_s, edges_t=edges_t, edge_attr=None, node_mask_s=atom_mask_s, 
-                    edge_mask_s=edge_mask_s, n_nodes_s=n_nodes_s, node_mask_t=atom_mask_t, edge_mask_t=edge_mask_t, 
-                    n_nodes_t=n_nodes_t, x_s=atom_positions_s, x_t=atom_positions_t)
-
-
-        loss = criterion(pred, label)  # Compute the loss.
-        epoch_loss += loss.item() * batch_size_s
+        
+        # pred = model(h0_s=one_hot_s, h0_t=one_hot_t, edges_s=edges_s, edges_t=edges_t, edge_attr=None, node_mask_s=atom_mask_s, 
+        #             edge_mask_s=edge_mask_s, n_nodes_s=n_nodes_s, node_mask_t=atom_mask_t, edge_mask_t=edge_mask_t, 
+        #             n_nodes_t=n_nodes_t, x_s=atom_positions_s, x_t=atom_positions_t)
+        pred = model(h0=[one_hot_s, one_hot_t], x=[atom_positions_s, atom_positions_t], all_edges=[edges_s, edges_t],
+                         all_edge_attr=[None, None], node_masks=[atom_mask_s, atom_mask_t],
+                         edge_masks=[edge_mask_s, edge_mask_t], n_nodes=[n_nodes_s, n_nodes_t])
+        loss = loss_func(pred, label)
+        epoch_loss += loss
         loss.backward()  # Derive gradients.
         optimizer.step()  # Update parameters based on gradients.
         optimizer.zero_grad()  # Clear gradients.
-
-        if i % 10 == 0:
-            print("Epoch %d \t Iteration %d \t loss %.4f" % (epoch, i, loss.item()))
     return epoch_loss
-    
+
 def test(loader):
     model.eval()
 
-    epoch_loss = 0
+    squared_error = []
+    error = 0 
     for data in loader:  # Iterate in batches over the training/test dataset.
         dense_positions_s, atom_mask_s = tg.utils.to_dense_batch(data.positions_s, data.positions_s_batch)
         dense_positions_t, atom_mask_t = tg.utils.to_dense_batch(data.positions_t, data.positions_t_batch)
@@ -127,55 +109,50 @@ def test(loader):
         batch_size_t, n_nodes_t, _ = dense_positions_t.size()
         atom_positions_s = dense_positions_s.view(batch_size_s * n_nodes_s, -1).to(device, dtype)
         atom_positions_t = dense_positions_t.view(batch_size_t * n_nodes_t, -1).to(device, dtype)
-
         edge_mask_s = atom_mask_s.unsqueeze(1) * atom_mask_s.unsqueeze(2)
-        #mask diagonal
         diag_mask = ~torch.eye(edge_mask_s.size(1), dtype=torch.bool).unsqueeze(0)
         edge_mask_s *= diag_mask
         edge_mask_s = edge_mask_s.view(batch_size_s * n_nodes_s * n_nodes_s, 1).to(device)
         edge_mask_t = atom_mask_t.unsqueeze(1) * atom_mask_t.unsqueeze(2)
-        #mask diagonal
         diag_mask = ~torch.eye(edge_mask_t.size(1), dtype=torch.bool).unsqueeze(0)
         edge_mask_t *= diag_mask
         edge_mask_t = edge_mask_t.view(batch_size_t * n_nodes_t * n_nodes_t, 1).to(device)
-
         atom_mask_s = atom_mask_s.view(batch_size_s * n_nodes_s, -1).to(device)
         atom_mask_t = atom_mask_t.view(batch_size_t * n_nodes_t, -1).to(device)
-
         one_hot_s, one_hot_s_mask = tg.utils.to_dense_batch(data.x_s, data.x_s_batch)
         one_hot_t, one_hot_s_mask = tg.utils.to_dense_batch(data.x_t, data.x_t_batch)
         one_hot_s = one_hot_s.view(batch_size_s * n_nodes_s, -1).to(device)
         one_hot_t = one_hot_t.view(batch_size_t * n_nodes_t, -1).to(device)
         edges_s = get_adj_matrix(n_nodes_s, batch_size_s, device)
         edges_t = get_adj_matrix(n_nodes_t, batch_size_t, device)
+        enviro = data.enviro.to(device, dtype)
         label = data.y.to(device, dtype)
-        pred = model(h0_s=one_hot_s, h0_t=one_hot_t, edges_s=edges_s, edges_t=edges_t, edge_attr=None, node_mask_s=atom_mask_s, 
-                    edge_mask_s=edge_mask_s, n_nodes_s=n_nodes_s, node_mask_t=atom_mask_t, edge_mask_t=edge_mask_t, 
-                    n_nodes_t=n_nodes_t, x_s=atom_positions_s, x_t=atom_positions_t)
+        
+        # pred = model(h0_s=one_hot_s, h0_t=one_hot_t, edges_s=edges_s, edges_t=edges_t, edge_attr=None, node_mask_s=atom_mask_s, 
+        #             edge_mask_s=edge_mask_s, n_nodes_s=n_nodes_s, node_mask_t=atom_mask_t, edge_mask_t=edge_mask_t, 
+        #             n_nodes_t=n_nodes_t, x_s=atom_positions_s, x_t=atom_positions_t)
+        pred = model(h0=[one_hot_s, one_hot_t], x=[atom_positions_s, atom_positions_t], all_edges=[edges_s, edges_t],
+                    all_edge_attr=[None, None], node_masks=[atom_mask_s, atom_mask_t],
+                    edge_masks=[edge_mask_s, edge_mask_t], n_nodes=[n_nodes_s, n_nodes_t], enviro=enviro)
+        loss = loss_func(pred, label)
+        error += loss.detach()
+    return error
 
-        # epoch_loss += criterion(pred, (label - prop_mean) / prop_mad).item()*batch_size
-        epoch_loss += criterion(pred, label).item()*batch_size_s
+training_loss = []
+testing_loss = []
 
-    return epoch_loss
+for epoch in range(1, n_epochs):
+    train()
+    train_acc = test(train_loader)
+    test_acc = test(test_loader)
+    training_loss.append(train_acc.detach().numpy())
+    testing_loss.append(test_acc.detach().numpy())
+    print(f'Epoch: {epoch:03d}, Train MSE: {train_acc:.4f}, Test MSE: {test_acc:.4f}')
 
-res = {'epochs': [], 'train_loss': [],'test_loss': [], 'best_val': 1e10, 'best_test': 1e10, 'best_epoch': 0}
-
-for epoch in range(0, n_epochs):
-    train_loss = train(epoch, train_loader)
-    res['train_loss'].append(train_loss)
-    if epoch % 4 == 0:
-        test_loss = test(test_loader)
-        res['epochs'].append(epoch)
-        res['test_loss'].append(test_loss)
-
-        if test_loss < res['best_val']:
-            res['best_val'] = test_loss
-            res['best_test'] = test_loss
-            res['best_epoch'] = epoch
-        print("test loss: %.4f \t epoch %d" % (test_loss, epoch))
-        print("Best: val loss: %.4f \t test loss: %.4f \t epoch %d" % (res['best_val'], res['best_test'], res['best_epoch']))
-
-
-    # json_object = json.dumps(res, indent=4)
-    # with open(args.outf + "/" + args.exp_name + "/losess.json", "w") as outfile:
-    #     outfile.write(json_object)
+plt.figure(figsize=(12,6))
+plt.plot(range(1, n_epochs), training_loss,label='training')
+plt.plot(range(1, n_epochs), testing_loss,label='testing')
+plt.ylabel('MSE')
+plt.xlabel('Epochs')
+plt.legend()
+plt.savefig('result.png')
